requirements.txt
streamlit
torch
torchvision
ultralytics
pillow
numpy
opencv-python-headless
matplotlib
